👷 #9: Build binaries on tags (#88)
* #9 Inited goreleaser

* #9: Inited the goreleaser

* #9 Added a GH Action to release binaries on tags

* #9: Setup release of binaries & announcement

* #9: fixed syntax errors

* #9 Commented out some lines that could cause noise & made the release workflow be dependant on the lint & vuln workflows

* #9: allow to run linting & vuln checks on tags

* #9 fixing dependencies between workflows

* #9 Tried to check the workflow name

* #9 Trying a different approach to reuse the other workflows in the release workflow

* #9 Disabled linux arm builds

* #9 disabled build for linux 386

* #9 removed s390x arch as it fails

* #9 removed the ppc64le arch because of failures

* #9: Removed riscv64 arch

* #9: Tried to fit a big number into uint32 on 32bit systems

* #9: fixed the cmd

* #9: fixed linting

* #9: Removed 386 as a arch target

* #9: ignored win arm64 pair

* #9: removed openbsd because of netpoll issues

* #9 Removed win from the OS target

* #9: try to force it release things

* #9: Fixed the token assignment

* #9: Added images and make commands to build them

* #9: Specified the linux as the target OS to fix issues with binary

* #9: Publishing the latest tags

* #9: completed the release workflows for images

* #9: disabled login into ghcr temporarily

* #9 commented more steps to speed up testing

* #9 exposing more info & change the working dir

* #9 using working directory to check out the right path

* #9: trying to fix the build time to keep the same in bins and images

* #9: Fixing the empty date

* #9 trying to fix the build time

* #9 fixing the ref

* #9 Trying to push images to GHCR

* #9 removed unneeded deps

* #9 Uncommented announcements on new releases

* #9: Passing Discord secrets to goreleaser

* #9: Added the basic homebrew tap setup

* #9 Made sure releases happen on the main branch only

* #9: Trying to use the deployment key to update the brew tap

+++ pkg/routers/health/buckets_test.go
@@ -10,7 +10,7 @@ import (
 
 func TestTokenBucket_Take(t *testing.T) {
 	bucketSize := 10
-	bucket := NewTokenBucket(1000, uint64(bucketSize))
+	bucket := NewTokenBucket(1000, uint(bucketSize))
 
 	for i := 0; i < bucketSize-1; i++ {
 		require.NoError(t, bucket.Take(1))
